Fix tests

+++ tests/run/assign.rs
@@ -23,7 +23,7 @@ fn inc(num: isize) -> isize {
 }
 
 #[no_mangle]
-extern "C" fn main(mut argc: i32, _argv: *const *const u8) -> i32 {
+extern "C" fn main(mut argc: isize, _argv: *const *const u8) -> i32 {
     argc = inc(argc);
     unsafe {
         libc::printf(b"%ld\n\0" as *const u8 as *const i8, argc);
